+++ inertia/components/projets/form/index.tsx
@@ -1,5 +1,15 @@
 import ParcellesStep from './parcelles-step'
 import ExploitationsStep from './exploitations-step'
 import CaptagesStep from './captages-step'
+import GeneralInfosStep from './general-infos-step'
+import ConsolidationsStep from './consolidations-step'
+import FirstEntryStep from './first-entry-step'
 
-export { ParcellesStep, ExploitationsStep, CaptagesStep }
+export {
+  ParcellesStep,
+  ExploitationsStep,
+  CaptagesStep,
+  GeneralInfosStep,
+  ConsolidationsStep,
+  FirstEntryStep,
+}
